apps/community_drf/requirements/base.txt
aiohttp==3.9.1
boto3==1.35.0
bump-my-version==0.22.0
celery==5.4.0
channels[daphne]==4.0.0
channels-redis==4.1.0
Django==4.2.11
dj-database-url==2.1.0
dj-rest-auth==4.0.1
django-allauth==0.50.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-environ==0.10.0
django-extensions==3.2.3
django-filter==23.2
django-redis==5.3.0
django-storages==1.14.4
django-rq==2.8.1
django-unfold==0.20.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.4
Markdown==3.4.4
pandas==2.2.2
Pillow==10.0.0
pymongo==4.5.0
pyTelegramBotAPI==4.18.1
psycopg[binary]==3.1.10
rq==1.15.1
